requirements.txt
youtube-dl==2021.12.17
vk-api==11.9.8
beautifulsoup4==4.11.1
pybooru==4.2.2
requests==2.28.0
rule34==1.8.1
discord==1.7.3
nest-asyncio==1.5.5
pandas==1.4.2
pymysql==1.0.2
python-dotenv==0.20.0
pygelbooru==0.5.0
sqlalchemy==2.0.17
asyncpg==0.27.0
psycopg2-binary==2.9.6